Column definitions removed from entity forms
